+++ src/pages/oganisation/department/DepartmentApi.js
@@ -1,9 +1,11 @@
 import axios from 'axios';
 
+const url = "http://localhost:8081"
+
 export const saveDepartment = async (formData) => {
     try{
         await axios.post(
-            "http://localhost:8081/department/create/department",
+            `${url}/department/create/department`,
             formData
           );
     } catch(error) {
@@ -13,7 +15,7 @@ export const saveDepartment = async (formData) => {
 
 export const deleteDepartment = async (id) => {
     try{
-        await axios.delete(`http://localhost:8081/department/delete/${id}`)
+        await axios.delete(`${url}/department/delete/${id}`)
     } catch(error) {
         console.error("Error deleting department",error)
     }
@@ -22,7 +24,7 @@ export const deleteDepartment = async (id) => {
 export const loadDepartment = async () => {
     try {
        const result =  await axios.get(
-            "http://localhost:8081/department/get/department",
+            `${url}/department/get/department`,
             {
               validateStatus: () => {
                 return true;
@@ -39,7 +41,7 @@ export const loadDepartment = async () => {
 export const fetchCompanies = async () => {
     try {
         const response = await axios.get(
-            "http://localhost:8081/company/get/company"
+            `${url}/company/get/company`
           );
           return response.data
     } catch (error){
@@ -51,10 +53,10 @@ export const fetchCompanies = async () => {
 export const fetchLocations = async () => {
     try {
         const response = await axios.get(
-          "http://localhost:8081/location/get/location"
+          `${url}/location/get/location`
         );
        return response.data 
       } catch (error) {
-        console.error("Error fetching department data", error);
+        console.error("Error fetching location data", error);
       }
 }
