+++ src/pages/oganisation/expenses/Mainfile/ExpensesProfile.js
@@ -1,0 +1,124 @@
+import React, { useEffect, useState } from "react";
+import { useParams } from "react-router-dom";
+import axios from "axios";
+import Header from "../../../../components/Header";
+import SideBar from "../../../../components/SideBar";
+
+const ExpensesProfile = () => {
+  const { id } = useParams();
+
+  const [expenses, setExpenses] = useState({
+    expenceType: "",
+    purchaseDate: "",
+    amount: "",
+    purchaseBy: "",
+    remarks: "",
+  });
+
+  useEffect(() => {
+    loadExpenses();
+  }, []);
+
+  const loadExpenses = async () => {
+    const result = await axios.get(`http://localhost:8081/expence/get/${id}`);
+    setExpenses(result.data);
+  };
+
+  return (
+    <div>
+      <Header />
+      <div className="dashboard-container">
+        <SideBar />
+        <div className="head-foot-part" style={{ padding: "0" }}>
+          <section className="shadow" style={{ backgroundColor: "whitesmoke" }}>
+            <div className="container py-5">
+              <div className="row">
+                <div className="col-lg-3">
+                  <div className="card mb-4">
+                    <div className="card-body text-center">
+                      <img
+                        src="https://mdbcdn.b-cdn.net/img/Photos/new-templates/bootstrap-chat/ava3.webp"
+                        alt="avatar"
+                        className="rounded-circle img-fluid"
+                        style={{ width: 150 }}
+                      />
+                    </div>
+                  </div>
+                </div>
+
+                <div className="col-lg-9">
+                  <div className="card mb-4">
+                    <div className="card-body">
+                      <hr />
+
+                      <div className="row">
+                        <div className="col-sm-3">
+                          <h5 className="mb-0">Expense Type</h5>
+                        </div>
+
+                        <div className="col-sm-9">
+                          <p className="text-muted mb-0">
+                            {expenses.expenceType}
+                          </p>
+                        </div>
+                      </div>
+
+                      <hr />
+
+                      <div className="row">
+                        <div className="col-sm-3">
+                          <h5 className="mb-0">Purchase Date</h5>
+                        </div>
+
+                        <div className="col-sm-9">
+                          <p className="text-muted mb-0">
+                            {expenses.purchaseDate}
+                          </p>
+                        </div>
+                      </div>
+                      <hr />
+
+                      <div className="row">
+                        <div className="col-sm-3">
+                          <h5 className="mb-0">Amount</h5>
+                        </div>
+
+                        <div className="col-sm-9">
+                          <p className="text-muted mb-0">{expenses.amount}</p>
+                        </div>
+                      </div>
+                      <hr />
+
+                      <div className="row">
+                        <div className="col-sm-3">
+                          <h5 className="mb-0">Purchased By</h5>
+                        </div>
+
+                        <div className="col-sm-9">
+                          <p className="text-muted mb-0">
+                            {expenses.purchaseBy}
+                          </p>
+                        </div>
+                      </div>
+                      <div className="row">
+                        <div className="col-sm-3">
+                          <h5 className="mb-0">Remarks</h5>
+                        </div>
+
+                        <div className="col-sm-9">
+                          <p className="text-muted mb-0">{expenses.remarks}</p>
+                        </div>
+                      </div>
+                    </div>
+                  </div>
+                </div>
+              </div>
+            </div>
+          </section>
+        </div>
+      </div>
+    </div>
+  );
+};
+
+export default ExpensesProfile;
